llm-evaluation-main/requirements.txt
requests
pandas
matplotlib
seaborn
python-dotenv
tqdm
streamlit
plotly
streamlit-ace
streamlit-option-menu
